Finalizing greatly improved dictionaries! - Dicts can now have Primi values as keys.   - Previously, due to the trivial implementation of dicts as ordinary PHP     arrays, the key always had to be a scalar value internally.     - For example false and 0 resolved into the same key before. Now they       are separate. - New internal structure MapContainer allowing to store key: value pairs   having Value objects as both keys and values. - Value::hash() function used to build dict keys (used in MapContainer).   - All types that can be used as dict keys must return a hash, other types     must throw UnhashableTypeException (this is default in the base Value     class) - Value::buildAutomatic() will return the same object, if it already was   an instance of Value. - ./primi script has new option "-ps" or "--print-stats" that   currently displays memory peak usage upon program exit.

+++ src/extensions/psl/ListExtension.php
@@ -76,7 +76,7 @@ class ListExtension extends Extension {
 	): Value {
 
 		// If the index is not found, this will return null.
-		$index = $list->protectedIndex((float) $index->value, false);
+		$index = $list->protectedIndex((int) $index->value, false);
 		if ($index === null) {
 			return $default ?? new NullValue;
 		}
@@ -105,7 +105,7 @@ class ListExtension extends Extension {
 
 		} else {
 
-			$index = $list->protectedIndex((float) $index->value);
+			$index = $list->protectedIndex((int) $index->value);
 			$popped = $list->value[$index];
 
 			// Remove the item and reindex the list.
